requirements.txt
#cx-Oracle==5.2
diff-match-patch==20121119
Django==1.9.6
django-annoying==0.10.3
django-object-actions==0.9.0
django-admin-extensions==0.10.0
django-registration-redux==1.6

django-reversion==2.0.8
django-reversion-compare==0.7.4

django-suit==0.2.25
djangorestframework==3.3.0
et-xmlfile==1.0.1
openpyxl==2.3.5
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
unicodecsv==0.14.1
django-forms-bootstrap==3.1.0

#heroku
dj-database-url==0.4.2
psycopg2==2.7.1
gunicorn==19.6.0
whitenoise==2.0.6

#cubes-viewer
argparse==1.2.1
requests==2.5.1
django-cors-headers==1.1.0

django-paypal==0.3.6

Django-Select2==5.10.0

# cubes
cubes==1.1
Flask==0.12.2
grako==3.99.9
Jinja2==2.9.6
SQLAlchemy==1.1.10
Werkzeug==0.12.2

# ?click==6.7
# ?defusedxml==0.5.0
# ?django-appconf==1.0.2
# ?django-pkgconf==0.2.1
# ?expressions==0.2.3
# ?html5lib==1.0b10
# ?httplib2==0.10.3
# ?icalendar==3.11.3
# ?itsdangerous==0.24
# ?jdcal==1.3
# ?jsonschema==2.6.0
# ?MarkupSafe==1.0
# ?oauthlib==2.0.2
# ?olefile==0.44
# ?pymongo==3.4.0
# ?PyPDF2==1.26.0
# ?python3-openid==3.1.0
# ?requests-oauthlib==0.8.0
# ?sqlparse==0.2.3
# ?webencodings==0.5.1
# ?xhtml2pdf==0.2b1
# ?xlwt==1.2.0
